Align unknown type handling with upstream (#9686)
In Postgres / Greenplum, a literal (e.g. NULL or 'text') in a query with
no obvious type gets a pseudo-type "unknown". A column that refers to
such literals will also get the "unknown" type.

We used to have logic to infer type of unknown literals in a subquery
based on the context. For example, the following query returns 124 as
foo is inferred to be an integer:

```sql
select foo + 123 from (select '1' as foo) a;
```

As another example, Greenplum infers the type of NULL as int in the
following "INSERT ... SELECT" with unknown-typed columns from a nested
subquery:

```sql
-- CREATE TABLE foo (a int, b text, c int);

INSERT INTO foo
SELECT * FROM (
  '1',
  'bbb',
  NULL
) bar(x,y,z);
```

While this user experience is nice, this cleverness is flawed.
Specifically, given an operator expression (let's say it's an addition)
where one operand has "unknown" type, and the other operand has type T
(let's say int), it's questionable that we can infer the unknown must be
a T (an int). In fact, the number of operators that fit the pattern of
"+(T, ?)" could be either ambiguous, or nonexistent.

Of course, there are scenarios where types are clearer: the SELECT list
that feeds into an INSERT or UPDATE has to be compatible (read:
cast-ible) with the target columns. Similarly, the SELECT list that goes
into a set operation (e.g. UNION) has to be compatible with
corresponding columns from other sub-queries. Upstream Postgres already
handles a large part of these cases.

This commit removes the Greenplum-specific cleverness. See the updated
tests for the behavior changes. Because this mostly changes
parse-analysis, expressions that are already stored in the catalog are
_not_ impacted (read: we will be able to restore the output of pg_dump).

Not that it will give us feature parity, but in Postgres 10, we will
coerce remaining unknown typed literals to text. See Postgres commit
1e7c4bb00497.

+++ src/test/regress/sql/bfv_joins.sql
@@ -227,8 +227,8 @@ select * from nlj1, nlj2 where nlj1.a = nlj2.a;
 explain select * from nlj1, nlj2 where nlj1.a is not distinct from nlj2.a;
 select * from nlj1, nlj2 where nlj1.a is not distinct from nlj2.a;
 
-explain select * from nlj1, (select NULL a, b from nlj2) other where nlj1.a is not distinct from other.a;
-select * from nlj1, (select NULL a, b from nlj2) other where nlj1.a is not distinct from other.a;
+explain select * from nlj1, (select NULL::int a, b from nlj2) other where nlj1.a is not distinct from other.a;
+select * from nlj1, (select NULL::int a, b from nlj2) other where nlj1.a is not distinct from other.a;
 
 explain select * from nlj1, nlj2 where nlj1.a is distinct from nlj2.a;
 select * from nlj1, nlj2 where nlj1.a is distinct from nlj2.a;
